requirements.txt
Django 
djangorestframework
django-cors-headers
dj-database-url
gunicorn
mysqlclient









